2885: Refactored getRelatives to EntityModel
