+++ packages/entity-server/src/routes/hierarchy/MultiEntityRelativesRoute.ts
@@ -4,16 +4,15 @@
  */
 
 import { Route } from '@tupaia/server-boilerplate';
-import { EntityType } from '../../../models';
-import { formatEntitiesForResponse } from '../format';
+import { EntityType } from '../../models';
+import { formatEntitiesForResponse } from './format';
 import {
   MultiEntityRequest,
   MultiEntityRequestParams,
   RequestBody,
   MultiEntityRequestQuery,
   EntityResponse,
-} from '../types';
-import { getMatchingRelatives } from './getMatchingRelatives';
+} from './types';
 
 export type MultiEntityRelativesRequest = MultiEntityRequest<
   MultiEntityRequestParams,
@@ -23,12 +22,12 @@ export type MultiEntityRelativesRequest = MultiEntityRequest<
 >;
 export class MultiEntityRelativesRoute extends Route<MultiEntityRelativesRequest> {
   async buildResponse() {
-    const { models, ctx } = this.req;
-    const { entities, fields, field } = ctx;
+    const { hierarchyId, entities, fields, field, filter } = this.req.ctx;
+
     const responseEntities: EntityType[] = [];
     await Promise.all(
       entities.map(async entity => {
-        const entitiesToUse = await getMatchingRelatives(models, { ...ctx, entity });
+        const entitiesToUse = await entity.getRelatives(hierarchyId, filter);
         responseEntities.push(...entitiesToUse);
       }),
     );
