--- packages/entity-server/src/routes/hierarchy/relatives/index.ts
@@ -1,10 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
- */
-
-export { EntityRelativesRoute, RelativesRequest } from './EntityRelativesRoute';
-export {
-  MultiEntityRelativesRoute,
-  MultiEntityRelativesRequest,
-} from './MultiEntityRelativesRoute';
